change: organize path structure
