+++ src/cli/main.lua
@@ -1,7 +1,7 @@
 local os = require('os')
-local zeebo_args = require('src/shared/args')
-local zeebo_meta = require('src/cli/meta')
-local zeebo_fs = require('src/cli/fs')
+local zeebo_args = require('src/lib/common/args')
+local zeebo_meta = require('src/lib/cli/meta')
+local zeebo_fs = require('src/lib/cli/fs')
 
 --! @cond
 local run = zeebo_args.has(arg, 'run')
@@ -15,57 +15,57 @@ local game = zeebo_args.param(arg, {'core', 'screen', 'dist'}, 2, '')
 
 local core_list = {
     repl={
-        src='src/lib/repl/main.lua',
-        exe='lua src/lib/repl/main.lua '..game,
+        src='src/engine/core/repl/main.lua',
+        exe='lua src/engine/core/repl/main.lua '..game,
         post_exe='lua dist/main.lua'
     },
     love={
-        src='src/lib/love2d/main.lua',
-        exe='love src/lib/love2d --screen '..screen..' '..game,
+        src='src/engine/core/love/main.lua',
+        exe='love src/engine/core/love --screen '..screen..' '..game,
         post_exe='love dist --screen '..screen
     },
     ginga={
-        src='src/lib/ginga/main.lua',
+        src='src/engine/core/ginga/main.lua',
         post_exe='ginga dist/main.ncl -s '..screen,
         extras={
-            'src/lib/ginga/main.ncl'
+            'src/engine/core/ginga/main.ncl'
         }
     },
     html5_webos={
-        src='src/lib/html5/main.lua',
+        src='src/engine/core/html5/main.lua',
         post_exe='webos24 $(pwd)/dist',
         pipeline={
             zeebo_meta.late(game):file(dist..'index.html'):file(dist..'appinfo.json'):pipe()
         },
         extras={
-            'src/lib/html5_webos/appinfo.json',
-            'src/lib/html5_webos/icon.png',
-            'src/lib/html5/index.html',
-            'src/lib/html5/index.html',
-            'src/lib/html5/engine.js',
+            'src/engine/meta/html5_webos/appinfo.json',
+            'src/engine/core/html5/index.html',
+            'src/engine/core/html5/index.html',
+            'src/engine/core/html5/engine.js',
+            'assets/icon80x80.png'
         }
     },
     html5_ginga={
-        src='src/lib/html5/main.lua',
+        src='src/engine/core/html5/main.lua',
         post_exe='ginga dist/main.ncl -s '..screen,
         pipeline={
             zeebo_meta.late(game):file(dist..'index.html'):pipe()
         },
         extras={
-            'src/lib/html5_ginga/main.ncl',
-            'src/lib/html5/index.html',
-            'src/lib/html5/index.html',
-            'src/lib/html5/engine.js',
+            'src/engine/meta/html5_ginga/main.ncl',
+            'src/engine/core/html5/index.html',
+            'src/engine/core/html5/index.html',
+            'src/engine/core/html5/engine.js',
         }
     },
     html5={
-        src='src/lib/html5/main.lua',
+        src='src/engine/core/html5/main.lua',
         pipeline={
             zeebo_meta.late(game):file(dist..'index.html'):pipe()
         },
         extras={
-            'src/lib/html5/index.html',
-            'src/lib/html5/engine.js'
+            'src/engine/core/html5/index.html',
+            'src/engine/core/html5/engine.js'
         }
     }
 }
