+++ src/lib/object/game.lua
@@ -1,4 +1,4 @@
---! @file src/object/game.lua
+--! @file src/lib/object/game.lua
 --! @short game object
 --! @brief a table to put anything related to developer game.
